+++ core/src/main/kotlin/fr/xgouchet/elmyr/annotation/RegexForgery.kt
@@ -1,0 +1,16 @@
+package fr.xgouchet.elmyr.annotation
+
+/**
+ * Mark a field, property or method parameter as a Regex based String forgery.
+ */
+@Target(
+        AnnotationTarget.VALUE_PARAMETER,
+        AnnotationTarget.FIELD,
+        AnnotationTarget.PROPERTY
+)
+@Retention(AnnotationRetention.RUNTIME)
+@MustBeDocumented
+annotation class RegexForgery(
+    /** The regex pattern to match. */
+    val value: String
+)